+++ database/migrations/2016_04_26_170951_add_experience_to_users_table.php
@@ -13,7 +13,7 @@ class AddExperienceToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('experience'); // 积分
+            $table->float('experience'); // 经验
         });
     }
